dev_hiertext/clean_requirements.txt
absl-py==2.3.1
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
albumentations==0.4.3
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
altair==4.2.2
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
aria2==0.0.1b0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.4.0
audioread==3.1.0
av==16.1.0
basicsr==1.4.2
bitarray==3.8.0
blinker==1.9.0
braceexpand==0.1.7
brotli==1.2.0
cachetools==7.0.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cmake==3.25.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
crcmod==1.7
cryptography==45.0.7
ctranslate2==4.7.1
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
cycler==0.12.1
Cython==0.29.33
datasets==2.14.5
decorator==5.2.1
dill==0.3.7
easydict==1.10
editdistance==0.8.1
einops==0.4.1
entrypoints==0.4
exceptiongroup==1.2.2
face-alignment==1.3.5
facexlib==0.3.0
fairseq==0.12.2
faiss-cpu==1.7.3
fastapi==0.88.0
faster-whisper==1.2.1
ffmpeg-python==0.2.0
ffmpy==0.3.1
filelock==3.20.3
filterpy==1.4.5
Flask==3.1.2
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2023.6.0
ftfy==6.3.1
future==1.0.0
gast==0.4.0
gfpgan==1.3.8
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.48.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gradio==3.50.0
gradio_client==0.6.1
groovy==0.1.2
grpcio==1.74.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.23.5
humanfriendly==10.0
hydra-core==1.0.7
hyper-connections==0.4.7
idna==3.11
imageio==2.9.0
imageio-ffmpeg==0.4.2
imgaug==0.2.6
importlib_metadata==8.7.1
importlib_resources==6.5.2
itsdangerous==2.2.0
jieba==0.42.1
Jinja2==3.1.6
jiter==0.12.0
jmespath==0.10.0
joblib==1.1.0
json5==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keras==2.13.1
kiwisolver==1.4.9
kornia==0.6.7
langchain==1.2.6
langchain-core==1.2.7
langchain-openai==1.1.7
langgraph==1.0.6
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.6
langgraph-sdk==0.3.3
langsmith==0.6.4
libclang==18.1.1
librosa==0.9.2
lightning-utilities==0.15.2
linkify-it-py==2.0.3
lit==15.0.7
llvmlite==0.39.0
lmdb==1.7.5
local-attention==1.11.2
loguru==0.7.3
lxml==6.0.2
Markdown==3.10
markdown-it-py==2.2.0
MarkupSafe==2.1.5
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
mock==5.2.0
modelscope==1.10.0
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.15
mypy_extensions==1.1.0
narwhals==2.16.0
networkx==3.4.2
numba==0.56.4
numpy==1.23.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
oauthlib==3.3.1
omegaconf==2.2.3
onnxruntime==1.23.2
onnxruntime-gpu==1.23.2
open-clip-torch==2.7.0
openai==2.15.0
opencv-python==4.7.0.72
opt_einsum==3.4.0
orjson==3.11.5
ormsgpack==1.12.2
oss2==2.19.1
pandas==1.5.3
Pillow==9.5.0
pooch==1.9.0
portalocker==3.2.0
praat-parselmouth==0.4.7
propcache==0.4.1
protobuf==3.20.3
psutil==7.1.3
pyarrow==12.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pycryptodome==3.23.0
pydantic==1.10.26
pydantic_core==2.41.5
pydeck==0.9.1
pyDeprecate==0.3.1
pydub==0.25.1
Pygments==2.19.2
Pympler==1.1
pyparsing==3.2.5
pyre-extensions==0.0.29
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pytorch-lightning==1.5.0
pytz==2025.2
PyWavelets==1.8.0
pyworld==0.3.2
PyYAML==6.0.3
qwen-vl-utils==0.0.14
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
resampy==0.3.1
rich==14.3.2
rpds-py==0.30.0
rsa==4.9.1
rvc-python==0.1.5
sacrebleu==2.6.0
sacremoses==0.0.53
safehttpx==0.1.7
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.10.1
segment-anything==1.0
semantic-version==2.10.0
semver==3.0.4
sentencepiece==0.1.99
simplejson==3.20.2
smmap==5.0.2
sortedcontainers==2.4.0
sounddevice==0.5.5
soundfile==0.13.1
starlette==0.22.0
streamlit==1.20.0
subword-nmt==0.3.8
sympy==1.14.0
tabulate==0.9.0
tb-nightly==2.21.0a20251023
tenacity==9.1.2
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.3.0
test_tube==0.7.5
threadpoolctl==3.6.0
tifffile==2025.5.10
tiktoken==0.12.0
timm==0.6.7
tokenizers==0.13.3
toml==0.10.2
toolz==1.1.0
torchaudio==2.0.2+cu118
torchcrepe==0.0.20
torchfcpe==0.0.4
torchmetrics==0.11.4
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
triton==2.0.0
typer==0.21.1
typer-slim==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.5.0
tzdata==2025.3
tzlocal==5.3.1
uc-micro-py==1.0.3
ujson==5.11.0
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.40.0
uvloop==0.22.1
validators==0.35.0
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.6.0
webdataset==0.2.5
webrtcvad==2.0.10
websockets==11.0.3
Werkzeug==3.1.4
wrapt==2.1.1
xformers==0.0.20
xxhash==3.6.0
yacs==0.1.8
yapf==0.32.0
yarl==1.22.0
zipp==3.23.0
